refactoring printElements into print(format::human)
