+++ include/dlaf/matrix/print_human.h
@@ -9,14 +9,26 @@
 //
 #pragma once
 
-/// @file
-
-#include "dlaf/matrix.h"
+#include "dlaf/common/index2d.h"
+#include "dlaf/common/range2d.h"
+#include "dlaf/matrix/matrix.h"
+#include "dlaf/matrix/tile.h"
+#include "dlaf/types.h"
 
 namespace dlaf {
 
+namespace format {
+struct human {};
+}
+
+namespace matrix {
+
+namespace internal {
+
+/// Print a matrix in a human readable format to standard output
+// TODO: now is printed to standard output, would it be better to save it into a file?
 template <class T>
-void printElements(Matrix<T, Device::CPU>& mat) {
+void print(format::human, Matrix<T, Device::CPU>& mat) {
   SizeType nrow = mat.size().rows();
   SizeType ncol = mat.size().cols();
   SizeType blockrow = mat.blockSize().rows();
@@ -40,4 +52,18 @@ void printElements(Matrix<T, Device::CPU>& mat) {
   std::cout << " " << std::endl;
 }
 
+/// Print a tile in a human readable format to standard output
+template <class T>
+void print(format::human, const Tile<T, Device::CPU>& tile) {
+  for (SizeType ii = 0; ii < tile.size().rows(); ++ii) {
+    for (SizeType jj = 0; jj < tile.size().cols(); ++jj) {
+      std::cout << std::setprecision(5) << tile({ii, jj}) << " ";
+    }
+    std::cout << " " << std::endl;
+  }
+  std::cout << " " << std::endl;
+}
+
+}
+}
 }
