+++ test/unit/CMakeLists.txt
@@ -40,12 +40,6 @@ if(DLAF_WITH_CUDA)
   )
 endif()
 
-DLAF_addTest(test_matrix_output
-  SOURCES test_matrix_output.cpp
-  LIBRARIES dlaf.core
-  USE_MAIN HPX
-)
-
 # Generic libraries
 add_subdirectory(common)
 add_subdirectory(communication)
