--- test/unit/test_matrix_output.cpp
@@ -1,69 +1,0 @@
-//
-// Distributed Linear Algebra with Future (DLAF)
-//
-// Copyright (c) 2018-2019, ETH Zurich
-// All rights reserved.
-//
-// Please, refer to the LICENSE file in the root directory.
-// SPDX-License-Identifier: BSD-3-Clause
-//
-
-#include "dlaf/matrix.h"
-#include "dlaf/matrix/copy.h"
-
-#include <vector>
-
-#include <gtest/gtest.h>
-#include <hpx/include/util.hpp>
-#include <hpx/local/future.hpp>
-
-#include "dlaf/matrix_output.h"
-#include "dlaf_test/matrix/util_matrix.h"
-#include "dlaf_test/matrix/util_matrix_blas.h"
-#include "dlaf_test/util_types.h"
-
-using namespace dlaf;
-using namespace dlaf::comm;
-using namespace dlaf::matrix;
-using namespace dlaf::matrix::test;
-using namespace dlaf_test;
-using namespace testing;
-
-template <typename Type>
-class MatrixOutputLocalTest : public ::testing::Test {};
-
-TYPED_TEST_SUITE(MatrixOutputLocalTest, MatrixElementTypes);
-
-struct TestSizes {
-  LocalElementSize size;
-  TileElementSize block_size;
-};
-const std::vector<TestSizes> sizes({
-    {{6, 6}, {2, 2}},
-    {{6, 6}, {3, 3}},
-    {{8, 8}, {3, 3}},
-});
-
-GlobalElementSize globalTestSize(const LocalElementSize& size) {
-  return {size.rows(), size.cols()};
-}
-
-TYPED_TEST(MatrixOutputLocalTest, printElements) {
-  using Type = float;
-  auto el = [](const GlobalElementIndex& index) {
-    SizeType i = index.row();
-    SizeType j = index.col();
-    return TypeUtilities<Type>::element(i + j, j - i);
-  };
-
-  for (const auto& sz : sizes) {
-    Matrix<Type, Device::CPU> mat(sz.size, sz.block_size);
-    EXPECT_EQ(Distribution(sz.size, sz.block_size), mat.distribution());
-
-    set(mat, el);
-
-    std::cout << "Matrix mat " << mat << std::endl;
-    std::cout << "Printing elements" << std::endl;
-    printElements(mat);
-  }
-}
